Drop CSRF-validation as it will not be used, since we currently only run the system locally, and if we will run it in the cloud, we will use token based validation - which makes CSRF validation superfluous. Copy Flyway-script to a separate folder in the image, so that the scripts can be read from the filesystem and not from the classpath - which does not work in a fatjar for Flyway versions > 11.12.0. Update build.yml. Update libraries. Use response oobjets to return data

+++ src/main/kotlin/org/simpleinvoice/server/repository/model/ProductORM.kt
@@ -15,7 +15,7 @@ object ProductTable : UUIDTable("product") {
     val quantity = integer("quantity")
     val price = double("price")
     val currency = varchar("currency", 255)
-    val taxPercentage = integer("tax_percentage")
+    val taxPercentage = double("tax_percentage")
     val totalPrice = double("total_price")
     val inactive = bool("inactive")
 }
